friday-backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
langchain==0.3.18
langgraph==0.2.73
langchain-core==0.3.35
langchain-community==0.3.17
langchain-openai==0.3.6
langchain-groq==0.2.4
langchain-ollama==0.2.3
python-dotenv==1.0.1
pydantic==2.10.6
tavily-python==0.5.1
duckduckgo-search==7.3.2
